% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_cells.r
\name{cluster_cells,SingleCellExperiment,ClusterLouvain-method}
\alias{cluster_cells,SingleCellExperiment,ClusterLouvain-method}
\title{cluster_cells}
\usage{
\S4method{cluster_cells}{SingleCellExperiment,ClusterLouvain}(
  x,
  method,
  reduction = "umap",
  label = "cell",
  cluster = "cluster",
  metric = "NMI",
  ...
)
}
\arguments{
\item{x}{a SingleCellExperiment object}

\item{method}{a}

\item{...}{Arguments passed to validate(...)}
}
\description{
cluster_cells
}
