% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.r
\name{cluster_cells}
\alias{cluster_cells}
\title{cluster_cells}
\usage{
cluster_cells(x, method, ...)
}
\arguments{
\item{x}{a data object}

\item{method}{a metric method object}

\item{...}{Other arguments}
}
\description{
The generic function of cluster_cells
}
