% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{metrics,SingleCellExperiment,MetricsARI-method}
\alias{metrics,SingleCellExperiment,MetricsARI-method}
\title{metrics}
\usage{
\S4method{metrics}{SingleCellExperiment,MetricsARI}(x, method, label = "cell_type", cluster = "cluster")
}
\arguments{
\item{x}{a SingleCellExperiment object}

\item{method}{a MetricsARI object}
}
\description{
metrics
}
