% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{metrics,SingleCellExperiment,MetricsGraphConnectivity-method}
\alias{metrics,SingleCellExperiment,MetricsGraphConnectivity-method}
\title{metrics}
\usage{
\S4method{metrics}{SingleCellExperiment,MetricsGraphConnectivity}(x, method, label = "cell_type", graph = "snn")
}
\arguments{
\item{x}{a SingleCellExperiment object}

\item{method}{a MetricsARI object}
}
\description{
Graph connectivity
}
