% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{metrics,SingleCellExperiment,MetricsKBET-method}
\alias{metrics,SingleCellExperiment,MetricsKBET-method}
\title{metrics}
\usage{
\S4method{metrics}{SingleCellExperiment,MetricsKBET}(x, method, label = "cell_type", reduction = "umap")
}
\arguments{
\item{x}{a SingleCellExperiment object}

\item{method}{a MetricsKBET object}
}
\description{
kBET
}
