% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{metrics,SingleCellExperiment,MetricsNMI-method}
\alias{metrics,SingleCellExperiment,MetricsNMI-method}
\title{metrics}
\usage{
\S4method{metrics}{SingleCellExperiment,MetricsNMI}(x, method, label = "cell_type", cluster = "cluster")
}
\arguments{
\item{x}{a SingleCellExperiment object}

\item{method}{a MetricsNMI object}
}
\description{
metrics
}
