% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.r
\name{metrics}
\alias{metrics}
\title{metrics}
\usage{
metrics(x, method, ...)
}
\arguments{
\item{x}{a data object}

\item{method}{a metric method object}

\item{...}{Other arguments}
}
\description{
The generic function of validate
}
