% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kBET.R
\name{run_kBET}
\alias{run_kBET}
\title{Run kBET scoring}
\usage{
run_kBET(x, batch_label)
}
\arguments{
\item{x}{SummarizedExperiment object containing an integrated single cell counts matrix}
}
\value{
returns a kBET score object
}
\description{
Run kBET scoring
}
