% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_params.r
\name{set_params,character-method}
\alias{set_params,character-method}
\title{set_params}
\usage{
\S4method{set_params}{character}(method, ...)
}
\arguments{
\item{method}{The method name}

\item{...}{Other arguments}
}
\description{
set_params
}
