% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.r
\name{set_params}
\alias{set_params}
\title{set_params}
\usage{
set_params(method, ...)
}
\arguments{
\item{method}{Method name}

\item{...}{Other arguments}
}
\description{
The generic function of set_params
}
