% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.r
\name{validate,SingleCellExperiment-method}
\alias{validate,SingleCellExperiment-method}
\title{validate}
\usage{
\S4method{validate}{SingleCellExperiment}(
  x,
  batch_field = "batch",
  label_field = "cell_type",
  reduction = "umap",
  ...
)
}
\arguments{
\item{a}{SingleCellExperiment object}
}
\description{
validate
}
